OMT-208 hotfix to delete dedrem on review submit
Since the dedrem is now computed on submit, rejecting a claim on review makes it so that the claim is not further processed on "process claim" and therefore there are lingering dedrem that are out of date. This change will just remove it and let the process claim restore it if appropriate.